feat(code): Rename menuInitiaton into Manager & inline documentation
