+++ Assets/Scripts/UI/MainMenuManager.cs.meta
@@ -1,2 +1,2 @@
 fileFormatVersion: 2
-guid: 5e70489eacb9ad4428c3202c1625beba
+guid: 3d1981cf71715ad49833554972314480
