--- Assets/Scripts/UI/MenuInitiator.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 3d1981cf71715ad49833554972314480
